update: 임시저장페이지  함수 분리
