+++ src/lib/drafts/delDrafts.ts
@@ -1,0 +1,7 @@
+import browserClient from "@/utils/supabase/client";
+
+export const delDrafts = async (delList: string[]) => {
+  const response = await browserClient.from("drafts").delete().in("id", delList);
+
+  return response;
+};
